Added backend, moved from menus.py, and more
Finally. I feel incompetent taking this long to finish something this simple.

Moved code from menus.py to pesterchum.py. It didn't fit the file seeing
as it's technically not a menu, but something that'll only show up on
init.

The image that shows up on the update window is in img/pchumbanner.png
Change it to your heart's content.

Three new files (not counting pchumbanner.png):
update.py, version_latest.py, and changelog.txt

update.py consists of all the logic behind fetching the information from
the repository. It also easily allows for changing the URLs used to get
the changelog.md and version files.
I didn't want to add libraries, so to make do, I had to use urllib.
Which means that it will NOT strip HTML from the downloaded file.
Try to provide it links to the raw github file you want it to fetch.

version_latest.py is where the version directly obtained from source is
stored. Used for comparing with version.py to see if there's an update
available, and to display said version on the update window, if there
an update.

changelog.txt holds the changelog directly pulled from source, allowing
for both the current and latest/source changelog to be stored and accessed.

All of this depends on the existence of CHANGELOG.md and version.py, and
on the idea that version.py will only be updated on full releases.

The only thing I haven't tested is whether the version detection system
works. I would, but after spending the last ten minutes trying to find out
why Pesterchum was hanging with no exceptions being raised or anything
showing up in the log at all, I got tired.
(I forgot to increment the variable of one (1) while loop.)

Feel free any code you might want to, in order to better adapt it.

CTRL + LSHIFT + ALT + Z will bring the update window up at any time,
regardless of whether there is one or not.
