+++ pesterchum.py
@@ -17,6 +17,7 @@ if os.path.dirname(sys.argv[0]):
 
 import ostools
 import pytwmn
+from update import gitfetch, pc
 
 from user_profile import (
     userConfig,
@@ -34,7 +35,6 @@ from menus import (
     PesterMemoList,
     LoadingScreen,
     AboutPesterchum,
-    UpdateAvailable,
     UpdatePesterchum,
     AddChumDialog,
 )
@@ -65,7 +65,7 @@ from scripts.services import SERVICES, CUSTOMBOTS, BOTNAMES, translate_nickserv_
 
 try:
     from PyQt6 import QtCore, QtGui, QtWidgets, QtMultimedia
-    from PyQt6.QtGui import QAction, QActionGroup
+    from PyQt6.QtGui import QAction, QActionGroup, QShortcut
 
     # Give Linux QtMultimedia warning.
     if ostools.isLinux():
@@ -78,7 +78,7 @@ try:
 except ImportError:
     print("PyQt5 fallback (pesterchum.py)")
     from PyQt5 import QtCore, QtGui, QtWidgets, QtMultimedia
-    from PyQt5.QtWidgets import QAction, QActionGroup
+    from PyQt5.QtWidgets import QAction, QActionGroup, QShortcut
 
 # Data directory
 ostools.validateDataDir()
@@ -1245,7 +1245,7 @@ class PesterWindow(MovingWindow):
         self.tabmemo = None
         self.shortcuts = {}
         self.aboutwindow = None
-        self.updateavailable = None
+        self.newversiondetected = None
         self.moods = None
         self.currentMoodIcon = None
 
@@ -1433,8 +1433,7 @@ class PesterWindow(MovingWindow):
         self.chanServAction = QAction(self.theme["main/menus/help/chanserv"], self)
         self.chanServAction.triggered.connect(self.loadChanServ)
         self.aboutAction = QAction(self.theme["main/menus/help/about"], self)
-        # self.aboutAction.triggered.connect(self.aboutwindow)
-        self.aboutAction.triggered.connect(self.updateAvailable)
+        self.aboutAction.triggered.connect(self.aboutPesterchum)
 
         # Because I can't expect all themes to have this included.
         # if self.theme.has_key("main/menus/help/reportbug"):
@@ -1554,6 +1553,50 @@ class PesterWindow(MovingWindow):
 
         self.chooseServerAskedToReset = False
         self.chooseServer()
+
+        # TODO: test!!!!!!!!!!!!!
+
+        # checks for updates and triggers the action AFTER everything important has loaded
+        
+        self.checkForUpdates = QAction("UPDATE", self)
+        
+        # atm, simply to make sure nothing breaks, what i'm gonna do is have it fetch both buildVersion
+        # variables, split them up into strings, then compare each index by size, left to right
+        # again. there's DEFINITELY a better way to do this. but i'm a very inexperienced programmer
+        # and it's a pretty important feature that hasn't been done yet
+
+        # if you, the reader, whoever you may be, have the patience to go thru my code and fix it, 
+        # be my guest
+        # -mal
+        
+
+        self.checkUpdateManually = QShortcut(
+            QtGui.QKeySequence("Ctrl+Shift+Alt+z"),
+            self
+        )
+        
+        build1 = gitfetch(pc.CURRENT_VERSION)
+        build2 = gitfetch(pc.LATEST_VERSION)
+
+
+        build1.replace('\"', '')
+        build2.replace('\"', '')
+
+        valueBuild1 = build1.split(".")
+        valueBuild2 = build2.split(".")
+
+        valCheck = 0
+
+        self.checkUpdateManually.activated.connect(self.updateAvailable)
+        
+        while valCheck <=2:
+            if valueBuild1[valCheck] < valueBuild2[valCheck]:
+                self.checkForUpdates.trigger()
+            valCheck+=1
+        
+
+
+        self.checkForUpdates.triggered.connect(self.updateAvailable)
 
         # Update RE bot used 2 b here but has now been moved to self.connected(), since this is too early (~lisanne)
 
@@ -3418,11 +3461,11 @@ class PesterWindow(MovingWindow):
     
     # TODO: setup backend for this schtick -mal
     def updateAvailable(self):
-        if self.updateavailable:
+        if self.newversiondetected:
             return
-        self.updateavailable = UpdateAvailable(self)
-        self.updateavailable.exec()
-        self.updateavailable = None
+        self.newversiondetected = UpdateAvailable(self)
+        self.newversiondetected.exec()
+        self.newversiondetected = None
 
     @QtCore.pyqtSlot()
     def loadCalsprite(self):
@@ -4383,6 +4426,204 @@ class MainProgram(QtCore.QObject):
 
     def run(self):
         sys.exit(self.app.exec())
+
+class UpdateAvailable(QtWidgets.QDialog):
+
+    
+    def update(self):
+        QtGui.QDesktopServices.openUrl(
+            QtCore.QUrl(
+                "https://github.com/Dpeta/pesterchum-alt-servers/",
+                QtCore.QUrl.ParsingMode.TolerantMode,
+            )
+        )
+      
+    def __init__(self, parent=None):
+        QtWidgets.QDialog.__init__(self, parent)
+        self.mainwindow = parent
+        self.setStyleSheet(self.mainwindow.theme["main/defaultwindow/style"])
+        self.title = QtWidgets.QLabel("UPD8????")
+        self.setModal(True)
+        self.setSizeGripEnabled(True)
+
+        # primary elements
+
+        mainContainer = QtWidgets.QHBoxLayout()
+        lefthandLayout = QtWidgets.QVBoxLayout()
+        righthandLayout = QtWidgets.QVBoxLayout()
+
+
+        # elements for lefthandLayout
+
+        versionBannerLayout = QtWidgets.QVBoxLayout()
+        versionContainerLayout = QtWidgets.QVBoxLayout()
+        versionDetailsLayout = QtWidgets.QHBoxLayout()
+        versionConstsLayout = QtWidgets.QVBoxLayout()
+        versionValuesLayout = QtWidgets.QVBoxLayout()
+        self.update_banner = QtWidgets.QLabel()
+        self.new_version_title = QtWidgets.QLabel("A new version of Pesterchum is available!")
+        self.const_currentversion = QtWidgets.QLabel("Current version:")
+        self.const_latestversion = QtWidgets.QLabel("Latest version:")
+        self.var_currentversion = QtWidgets.QLabel()
+        self.var_latestversion = QtWidgets.QLabel()
+        self.acceptUpdate = QtWidgets.QPushButton("Get latest version!", self)
+
+
+        # elements for righthandLayout
+
+        scrollBoxContainer = QtWidgets.QHBoxLayout()
+        changelogContainer = QtWidgets.QVBoxLayout()
+        self.frame = QtWidgets.QFrame()
+        frameLayout = QtWidgets.QVBoxLayout(self.frame)
+        self.changelogWidgetContents = QtWidgets.QWidget()
+        changelogWidgetContentsLayout = QtWidgets.QVBoxLayout(self.changelogWidgetContents)
+        self.changelogScrollable = QtWidgets.QScrollArea()
+        self.changelogTitle = QtWidgets.QLabel("Changelog:")
+        self.changelogContents = QtWidgets.QLabel()
+        self.postpone = QtWidgets.QPushButton("Remind me later", self)
+
+
+        # size policies
+
+        spMaxMin = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Maximum, 
+            QtWidgets.QSizePolicy.Policy.Minimum
+            )
+        spMax = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Maximum, 
+            QtWidgets.QSizePolicy.Policy.Maximum
+            )
+        spMinExp = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.MinimumExpanding, 
+            QtWidgets.QSizePolicy.Policy.MinimumExpanding
+            )
+        spPref = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Preferred, 
+            QtWidgets.QSizePolicy.Policy.Preferred
+            )
+        spPrefMin = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Preferred, 
+            QtWidgets.QSizePolicy.Policy.Minimum
+            )
+        spButtons = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Policy.Minimum, 
+            QtWidgets.QSizePolicy.Policy.Fixed
+            )
+        
+
+        # since i'm overriding the style sheet, and i don't know exactly what that's gonna do,
+        # i'm giving it all these attributes as a fallback.
+        # TODO: replace this with actual, proper, thought-out code
+
+        title1 = QtGui.QFont()
+        title1.setFamily("Courier New")
+        title1.setBold(True)
+        title1.setUnderline(True)
+        title1.setItalic(True)
+
+        subtitle1 = QtGui.QFont()
+        subtitle1.setFamily("Courier New")
+        subtitle1.setBold(True)
+        subtitle1.setUnderline(False)
+        subtitle1.setItalic(False)
+        
+        subtitle2 = QtGui.QFont()
+        subtitle2.setFamily("Courier New")
+        subtitle2.setBold(False)
+        subtitle2.setUnderline(False)
+        subtitle2.setItalic(False)
+
+
+        # lefthand element configurations
+
+        self.var_currentversion.setFont(subtitle2)
+        self.var_currentversion.setSizePolicy(spMaxMin)
+        self.var_currentversion.setAlignment(QtCore.Qt.AlignmentFlag.AlignBottom|QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft)
+        self.var_currentversion.setText(gitfetch(pc.CURRENT_VERSION))
+
+        self.var_latestversion.setFont(subtitle2)
+        self.var_latestversion.setSizePolicy(spMaxMin)
+        self.var_latestversion.setAlignment(QtCore.Qt.AlignmentFlag.AlignBottom|QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft)
+        self.var_latestversion.setText(gitfetch(pc.LATEST_VERSION))
+
+        self.const_currentversion.setSizePolicy(spMax)
+        self.const_currentversion.setAlignment(QtCore.Qt.AlignmentFlag.AlignBottom|QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft)
+
+        self.const_latestversion.setSizePolicy(spMax)
+        self.const_latestversion.setAlignment(QtCore.Qt.AlignmentFlag.AlignBottom|QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft)
+
+        self.new_version_title.setAlignment(QtCore.Qt.AlignmentFlag.AlignBottom|QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignCenter)
+        self.new_version_title.setSizePolicy(spPrefMin)
+        self.new_version_title.setFont(title1)
+
+        self.update_banner.setSizePolicy(spMinExp)
+        self.update_banner.setText("<html><head/><body><p align=\"center\"><img src=\"img/pchumbanner.png\"/></p></body></html>")
+        
+        versionContainerLayout.setSizeConstraint(QtWidgets.QLayout.SizeConstraint.SetMaximumSize)
+        versionDetailsLayout.setSizeConstraint(QtWidgets.QLayout.SizeConstraint.SetMaximumSize)
+        versionConstsLayout.setSizeConstraint(QtWidgets.QLayout.SizeConstraint.SetMinimumSize)
+        versionValuesLayout.setSizeConstraint(QtWidgets.QLayout.SizeConstraint.SetMaximumSize)
+
+        self.acceptUpdate.setSizePolicy(spButtons)
+        self.acceptUpdate.clicked.connect(self.update)
+
+        # righthand element configurations
+
+        self.changelogTitle.setFont(subtitle1)
+        self.changelogTitle.setSizePolicy(spPref)
+        self.changelogContents.setText(gitfetch(pc.CHANGELOG))
+        self.changelogContents.setFont(subtitle2)
+        self.changelogContents.setSizePolicy(spMinExp)
+        self.changelogContents.setAlignment(QtCore.Qt.AlignmentFlag.AlignLeading|QtCore.Qt.AlignmentFlag.AlignLeft|QtCore.Qt.AlignmentFlag.AlignTop)
+        self.changelogWidgetContents.setGeometry(QtCore.QRect(0, 0, 173, 193))
+        self.changelogWidgetContents.setSizePolicy(spMinExp)
+
+        self.frame.setFrameShape(QtWidgets.QFrame.Shape.StyledPanel)
+        self.frame.setFrameShadow(QtWidgets.QFrame.Shadow.Raised)
+
+        self.changelogScrollable.setWidgetResizable(True)
+        self.changelogScrollable.setWidget(self.changelogWidgetContents)
+
+        self.postpone.clicked.connect(self.reject)
+
+
+        # nesting
+
+        mainContainer.addLayout(lefthandLayout)
+        mainContainer.addLayout(righthandLayout)
+
+        lefthandLayout.addLayout(versionBannerLayout)
+        lefthandLayout.addWidget(self.acceptUpdate)
+        
+        versionBannerLayout.addWidget(self.update_banner)
+        versionBannerLayout.addLayout(versionContainerLayout)
+
+        versionContainerLayout.addWidget(self.new_version_title)
+        versionContainerLayout.addLayout(versionDetailsLayout)
+
+        versionDetailsLayout.addLayout(versionConstsLayout)
+        versionDetailsLayout.addLayout(versionValuesLayout)
+
+        versionConstsLayout.addWidget(self.const_currentversion)
+        versionConstsLayout.addWidget(self.const_latestversion)
+
+        versionValuesLayout.addWidget(self.var_currentversion)
+        versionValuesLayout.addWidget(self.var_latestversion)
+        
+        righthandLayout.addWidget(self.frame)
+        righthandLayout.addWidget(self.postpone)
+
+        frameLayout.addLayout(scrollBoxContainer)
+        
+        scrollBoxContainer.addLayout(changelogContainer)
+
+        changelogContainer.addWidget(self.changelogTitle)
+        changelogContainer.addWidget(self.changelogScrollable)
+
+        self.changelogScrollable.setWidget(self.changelogWidgetContents)
+        changelogWidgetContentsLayout.addWidget(self.changelogContents)
+
+        self.setLayout(mainContainer)
 
 
 if __name__ == "__main__":
