+++ version_latest.py
@@ -1,0 +1,2 @@
+_pcVersion = "Alt. v2.6.0"
+buildVersion = "v2.6.0"
